requirements.txt
spacy==3.7.2
spacy-transformers==1.3.8
fastapi==0.115.0
uvicorn==0.30.6
pymupdf==1.24.10
nameparser==1.1.3
pytesseract==0.3.13
pillow==10.4.0
torch==2.0.1